distância e velocidade atualizadas no painel

+++ SkiFree/js/entities/tree.js
@@ -10,15 +10,15 @@ class Tree {
 
 	update(skierSpeed) {
 		if (!this.mustBeDrawn) return;
-		if (parseInt(this.element.style.top) < -this.element.clientHeight) {
+		if (parseFloat(this.element.style.top) < -this.element.clientHeight) {
 			this.mustBeDrawn = false;
 		}
 		
-		this.element.style.top = parseInt(this.element.style.top) - skierSpeed + 'px';
+		this.element.style.top = parseFloat(this.element.style.top) - skierSpeed + 'px';
 	}
 
 	checkCollisionWithSkier(skier) {
-		let rectA = new Rectangle(parseInt(this.element.style.top), parseInt(this.element.style.left), this.element.clientWidth, this.element.clientHeight);
+		let rectA = new Rectangle(parseFloat(this.element.style.top), parseFloat(this.element.style.left), this.element.clientWidth, this.element.clientHeight);
 		let rectB = new Rectangle(skier.top, skier.left, skier.element.clientWidth, skier.element.clientHeight);
 		
 		if (rectA.x1 < rectB.x2 && rectA.x2 > rectB.x1 && rectA.y1 < rectB.y2 && rectA.y2 > rectB.y1) {
